[MOVE-ONLY] Move CBudgetManager to budgetmanager.h/cpp
