+++ src/budget/budgetproposal.cpp
@@ -351,97 +351,3 @@ void CBudgetProposal::Relay()
     g_connman->RelayInv(inv);
 }
 
-
-/*
- * Check Collateral
- */
-bool CheckCollateralConfs(const uint256& nTxCollateralHash, int nCurrentHeight, int nProposalHeight, std::string& strError)
-{
-    const int nRequiredConfs = Params().GetConsensus().nBudgetFeeConfirmations;
-    const int nConf = nCurrentHeight - nProposalHeight + 1;
-
-    if (nConf < nRequiredConfs) {
-        strError = strprintf("Collateral requires at least %d confirmations - %d confirmations (current height: %d, fee tx height: %d)",
-                nRequiredConfs, nConf, nCurrentHeight, nProposalHeight);
-        LogPrint(BCLog::MNBUDGET,"%s: %s\n", __func__, strError);
-        return false;
-    }
-    return true;
-}
-
-bool CheckCollateral(const uint256& nTxCollateralHash, const uint256& nExpectedHash, std::string& strError, int64_t& nTime, int nCurrentHeight, bool fBudgetFinalization)
-{
-    CTransaction txCollateral;
-    uint256 nBlockHash;
-    if (!GetTransaction(nTxCollateralHash, txCollateral, nBlockHash, true)) {
-        strError = strprintf("Can't find collateral tx %s", nTxCollateralHash.ToString());
-        return false;
-    }
-
-    if (txCollateral.vout.size() < 1) return false;
-    if (txCollateral.nLockTime != 0) return false;
-
-    CScript findScript;
-    findScript << OP_RETURN << ToByteVector(nExpectedHash);
-
-    bool foundOpReturn = false;
-    for (const CTxOut &o : txCollateral.vout) {
-        if (!o.scriptPubKey.IsNormalPaymentScript() && !o.scriptPubKey.IsUnspendable()) {
-            strError = strprintf("Invalid Script %s", txCollateral.ToString());
-            return false;
-        }
-        if (fBudgetFinalization) {
-            // Collateral for budget finalization
-            // Note: there are still old valid budgets out there, but the check for the new 5 PIV finalization collateral
-            //       will also cover the old 50 PIV finalization collateral.
-            LogPrint(BCLog::MNBUDGET, "Final Budget: o.scriptPubKey(%s) == findScript(%s) ?\n", HexStr(o.scriptPubKey), HexStr(findScript));
-            if (o.scriptPubKey == findScript) {
-                LogPrint(BCLog::MNBUDGET, "Final Budget: o.nValue(%ld) >= BUDGET_FEE_TX(%ld) ?\n", o.nValue, BUDGET_FEE_TX);
-                if(o.nValue >= BUDGET_FEE_TX) {
-                    foundOpReturn = true;
-                }
-            }
-        } else {
-            // Collateral for normal budget proposal
-            LogPrint(BCLog::MNBUDGET, "Normal Budget: o.scriptPubKey(%s) == findScript(%s) ?\n", HexStr(o.scriptPubKey), HexStr(findScript));
-            if (o.scriptPubKey == findScript) {
-                LogPrint(BCLog::MNBUDGET, "Normal Budget: o.nValue(%ld) >= PROPOSAL_FEE_TX(%ld) ?\n", o.nValue, PROPOSAL_FEE_TX);
-                if(o.nValue >= PROPOSAL_FEE_TX) {
-                    foundOpReturn = true;
-                }
-            }
-        }
-    }
-
-    if (!foundOpReturn) {
-        strError = strprintf("Couldn't find opReturn %s in %s", nExpectedHash.ToString(), txCollateral.ToString());
-        return false;
-    }
-
-    // Retrieve block height (checking that it's in the active chain) and time
-    // both get set in CBudgetProposal/CFinalizedBudget by the caller (AddProposal/AddFinalizedBudget)
-    if (nBlockHash.IsNull()) {
-        strError = strprintf("Collateral transaction %s is unconfirmed", nTxCollateralHash.ToString());
-        return false;
-    }
-    nTime = 0;
-    int nProposalHeight = 0;
-    {
-        LOCK(cs_main);
-        BlockMap::iterator mi = mapBlockIndex.find(nBlockHash);
-        if (mi != mapBlockIndex.end() && (*mi).second) {
-            CBlockIndex* pindex = (*mi).second;
-            if (chainActive.Contains(pindex)) {
-                nProposalHeight = pindex->nHeight;
-                nTime = pindex->nTime;
-            }
-        }
-    }
-
-    if (!nProposalHeight) {
-        strError = strprintf("Collateral transaction %s not in Active chain", nTxCollateralHash.ToString());
-        return false;
-    }
-
-    return CheckCollateralConfs(nTxCollateralHash, nCurrentHeight, nProposalHeight, strError);
-}
