+++ src/budget/budgetproposal.h
@@ -15,9 +15,6 @@ static const CAmount BUDGET_FEE_TX_OLD = (50 * COIN);
 static const CAmount BUDGET_FEE_TX = (5 * COIN);
 static const int64_t BUDGET_VOTE_UPDATE_MIN = 60 * 60;
 
-// Used to check both proposals and finalized-budgets collateral txes
-bool CheckCollateral(const uint256& nTxCollateralHash, const uint256& nExpectedHash, std::string& strError, int64_t& nTime, int nCurrentHeight, bool fBudgetFinalization);
-
 //
 // Budget Proposal : Contains the masternode votes for each budget
 //
